updating authz checks in authorization group service.

+++ lib/service/group-service.ts
@@ -4,11 +4,14 @@ import TenantDao from "@/lib/dao/tenant-dao";
 import { GraphQLError } from "graphql/error/GraphQLError";
 import { randomUUID } from 'crypto'; 
 import GroupDao from "../dao/authorization-group-dao";
-import { NAME_ORDER_EASTERN, SEARCH_INDEX_OBJECT_SEARCH, SEARCH_INDEX_REL_SEARCH } from "@/utils/consts";
+import { AUTHORIZATION_GROUP_CREATE_SCOPE, AUTHORIZATION_GROUP_DELETE_SCOPE, AUTHORIZATION_GROUP_READ_SCOPE, AUTHORIZATION_GROUP_UPDATE_SCOPE, AUTHORIZATION_GROUP_USER_ASSIGN_SCOPE, AUTHORIZATION_GROUP_USER_REMOVE_SCOPE, NAME_ORDER_EASTERN, SEARCH_INDEX_OBJECT_SEARCH, SEARCH_INDEX_REL_SEARCH, TENANT_READ_ALL_SCOPE } from "@/utils/consts";
 import { getOpenSearchClient } from "@/lib/data-sources/search";
 import { Client } from "@opensearch-project/opensearch";
 import { DaoFactory } from "../data-sources/dao-factory";
 import IdentityDao from "../dao/identity-dao";
+import { authorizeCreateObject, authorizeDeleteObject, authorizeRead, authorizeUpdateObject, containsScope, filterResults, readWithInputFilterAndAuthorization } from "@/utils/authz-utils";
+import { AUTHORIZATION_GROUP_USER_ADD_MUTATION } from "@/graphql/mutations/oidc-mutations";
+import { Author } from "next/dist/lib/metadata/types/metadata-types";
 
 
 const tenantDao: TenantDao = DaoFactory.getInstance().getTenantDao();
@@ -25,15 +28,65 @@ class GroupService {
     }
 
     public async getGroups(tenantId?: string): Promise<Array<AuthorizationGroup>> {
-        return groupDao.getAuthorizationGroups(tenantId);
+
+        const getData = readWithInputFilterAndAuthorization(
+            groupDao.getAuthorizationGroups,
+            {
+                preProcess: async (oidcContext, tenantId?: string) => {
+                    if(!tenantId){
+                        if(oidcContext.portalUserProfile?.managementAccessTenantId !== oidcContext.rootTenant.tenantId){
+                            if(oidcContext.portalUserProfile && oidcContext.portalUserProfile.managementAccessTenantId){
+                                return [oidcContext.portalUserProfile.managementAccessTenantId];
+                            }
+                            else{
+                                return [tenantId];
+                            }
+                        }
+                        else{
+                            return [tenantId];
+                        }
+                    }
+                    return [tenantId];
+                },
+                authorize: async function (oidcContext: OIDCContext, tenantId?: string): Promise<{ isAuthorized: boolean; errorMessage: string | null; result: Array<AuthorizationGroup>}> {                    
+                    const groups = await groupDao.getAuthorizationGroups(tenantId);
+                    return {isAuthorized: true, errorMessage: "", result: groups}                    
+                }
+            }
+        );
+        const data = await getData(this.oidcContext, [TENANT_READ_ALL_SCOPE, AUTHORIZATION_GROUP_READ_SCOPE], tenantId);
+        return data ? data : [];
+        
     }
 
-    public async getGroupById(groupId: string): Promise<AuthorizationGroup | null> {        
-        const group: AuthorizationGroup | null = await groupDao.getAuthorizationGroupById(groupId);        
-        return Promise.resolve(group);        
+    public async getGroupById(groupId: string): Promise<AuthorizationGroup | null> {    
+        
+        const getData = readWithInputFilterAndAuthorization(
+            groupDao.getAuthorizationGroupById,
+            {
+                authorize: async function (oidcContext, groupId): Promise<{ isAuthorized: boolean; errorMessage: string | null; result: AuthorizationGroup | null}> {
+                    const result = await groupDao.getAuthorizationGroupById(groupId);
+                    if(result && result.tenantId !== oidcContext.portalUserProfile?.managementAccessTenantId){
+                        return {isAuthorized: false, errorMessage: "ERROR_INSUFFICIENT_PERMISSIONS_TO_READ_OBJECT", result: null};
+                    }
+                    else{
+                        return { isAuthorized: true, errorMessage: null, result: result};
+                    } 
+                }
+            }
+        );
+
+        const t = await getData(this.oidcContext, [AUTHORIZATION_GROUP_READ_SCOPE, TENANT_READ_ALL_SCOPE], groupId);
+        return t;           
     }
 
     public async createGroup(group: AuthorizationGroup): Promise<AuthorizationGroup> {
+
+        const {isAuthorized, errorMessage} = authorizeCreateObject(this.oidcContext, AUTHORIZATION_GROUP_CREATE_SCOPE, group.tenantId);
+        if(!isAuthorized){
+            throw new GraphQLError(errorMessage || "ERROR");
+        }
+
         const tenant: Tenant | null = await tenantDao.getTenantById(group.tenantId);
         if(!tenant){
             throw new GraphQLError("ERROR_TENANT_NOT_FOUND_FOR_GROUP_CREATION");
@@ -46,6 +99,12 @@ class GroupService {
     }
 
     public async updateGroup(group: AuthorizationGroup): Promise<AuthorizationGroup> {
+
+        const {isAuthorized, errorMessage} = authorizeUpdateObject(this.oidcContext, AUTHORIZATION_GROUP_UPDATE_SCOPE, group.tenantId);
+        if(!isAuthorized){
+            throw new GraphQLError(errorMessage || "ERROR");
+        }
+
         const existingGroup: AuthorizationGroup | null = await groupDao.getAuthorizationGroupById(group.groupId);
         if(!existingGroup){
             throw new GraphQLError("ERROR_GROUP_NOT_FOUND");
@@ -95,6 +154,12 @@ class GroupService {
     }
 
     public async deleteGroup(groupId: string): Promise<void> {
+
+        const {isAuthorized, errorMessage} = authorizeDeleteObject(this.oidcContext, AUTHORIZATION_GROUP_DELETE_SCOPE, groupId);
+        if(!isAuthorized){
+            throw new GraphQLError(errorMessage || "ERROR");
+        }
+
         throw new Error("Method not implemented.");
     }
 
@@ -114,6 +179,12 @@ class GroupService {
         const userTenantRel: UserTenantRel | null = await identityDao.getUserTenantRel(authzGroup.tenantId, userId);
         if(!userTenantRel){
             throw new GraphQLError("ERROR_INVALID_TENANT_FOR_USER");
+        }
+
+        // Is the user authorized to perform the action?
+        const {isAuthorized, errorMessage} = authorizeCreateObject(this.oidcContext, AUTHORIZATION_GROUP_USER_ASSIGN_SCOPE, authzGroup.tenantId);
+        if(!isAuthorized){
+            throw new GraphQLError(errorMessage || "ERROR");
         }
 
         const r: AuthorizationGroupUserRel = await groupDao.addUserToAuthorizationGroup(userId, groupId);
@@ -139,17 +210,29 @@ class GroupService {
     }
     
     public async removeUserFromGroup(userId: string, groupId: string): Promise<void> {
-        await groupDao.removeUserFromAuthorizationGroup(userId, groupId);
-        
-        await searchClient.delete({
-            id: `${groupId}::${userId}`,
-            index: SEARCH_INDEX_REL_SEARCH,
-            refresh: "wait_for"
-        });  
+        const authzGroup: AuthorizationGroup | null = await groupDao.getAuthorizationGroupById(groupId);
+        if(authzGroup){
+            const {isAuthorized, errorMessage} = authorizeDeleteObject(this.oidcContext, AUTHORIZATION_GROUP_USER_REMOVE_SCOPE, groupId);
+            if(!isAuthorized){
+                throw new GraphQLError(errorMessage || "ERROR");
+            }
+
+            await groupDao.removeUserFromAuthorizationGroup(userId, groupId);        
+            await searchClient.delete({
+                id: `${groupId}::${userId}`,
+                index: SEARCH_INDEX_REL_SEARCH,
+                refresh: "wait_for"
+            });  
+        }
+        return Promise.resolve();
     }
 
     public async getUserAuthorizationGroups(userId: string): Promise<Array<AuthorizationGroup>> {
-        return groupDao.getUserAuthorizationGroups(userId);
+        if(!containsScope([AUTHORIZATION_GROUP_READ_SCOPE, TENANT_READ_ALL_SCOPE], this.oidcContext.portalUserProfile?.scope || [])){
+            throw new GraphQLError("ERROR_NO_PERMISSION");
+        }
+        const groups: Array<AuthorizationGroup> = await groupDao.getUserAuthorizationGroups(userId);
+        return filterResults(groups, this.oidcContext, (g: AuthorizationGroup) => g.tenantId);        
     }
 }
